+++ src/utils.ts
@@ -1,12 +1,13 @@
-import { Uri, Webview } from "vscode";
+import { Uri, type Webview } from 'vscode'
 
-export function getNonce() {
-    let text = '';
-    const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    for (let i = 0; i < 32; i++) {
-        text += possible.charAt(Math.floor(Math.random() * possible.length));
-    }
-    return text;
+export function getNonce (): string {
+  let text = ''
+  const possible =
+    'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
+  for (let i = 0; i < 32; i++) {
+    text += possible.charAt(Math.floor(Math.random() * possible.length))
+  }
+  return text
 }
 
 /**
@@ -20,6 +21,10 @@ export function getNonce() {
  * @param pathList An array of strings representing the path to a file/resource
  * @returns A URI pointing to the file/resource
  */
-export function getUri(webview: Webview, extensionUri: Uri, pathList: string[]) {
-    return webview.asWebviewUri(Uri.joinPath(extensionUri, ...pathList));
+export function getUri (
+  webview: Webview,
+  extensionUri: Uri,
+  pathList: string[]
+): Uri {
+  return webview.asWebviewUri(Uri.joinPath(extensionUri, ...pathList))
 }